core/services/rag-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pinecone==3.0.0
google-cloud-aiplatform==1.38.0
google-cloud-secret-manager==2.16.4
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
